[c++ build] suppress deprecation warnings
With newer version of Xcode (at least starting with version 13.2.1),
building Kudu on macOS produces many warnings because of C++17-related
deprecations, so it's not easy to spot warnings attributed to updates
in the local workspace.  This patch addresses the issue, adding pragmas
to disable deprecated declarations where appropriate.

In addition, this patch also silences warnings related to usage of
the deprecated Kudu API in src/kudu/client/schema.cc (it's used just
in the library implementation, so it's not a big deal).  I also took the
liberty to replace include guards with 'pragma once' and address
some of clang-tidy's recommendations in a few affected files.

Change-Id: I069a0680b4bc981b311374cc08bb950578db0dd0
Reviewed-on: http://gerrit.cloudera.org:8080/19202
Tested-by: Kudu Jenkins
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>
Reviewed-by: Attila Bukor <abukor@apache.org>