[KUDU-3419] Fix the stuck of starting tablet server
If the permission of tablet metadata file is denied when
tablet server starting,tablet server will get stuck and
can not exit automatically. That is because if the permission
is denied, the object of TabletServer will be deconstructed,
it will deconstruct its thread pool:txn_status_manager_pool_,
but this thread pool contains a task: TxnStalenessTrackerTask,
the task is still running. There are no code to shutdown the
task in that case. Therefore tablet server will get stuck.

See #KUDU-3419 for detail

Change-Id: I8c9a4f4158fcb0a36499345e00ee72c65f5fefe0
Reviewed-on: http://gerrit.cloudera.org:8080/19203
Reviewed-by: Yingchun Lai <acelyc1112009@gmail.com>
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>