KUDU-3421 Specify the error causing CLI argument
The new error message helps to locate the exact word that caused the
problem by printing the first argument that is not a flag in
applications that only use flag arguments. This can happen if the user
forgets to put "-" to one of the arguments or incorrectly tries to use
an argument in space separated format.

Change-Id: I4d832615d4b70718ce581ab72a27bad012d6a87f
Reviewed-on: http://gerrit.cloudera.org:8080/19253
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>