KUDU-3423 Add support for Ranger KMS HA
Ranger KMS supports high availability by listing multiple Ranger KMS
servers (comma-separated, no spaces). Up until now, Kudu allowed
configuring only one Ranger KMS server.

This patch adds support for listing multiple KMS servers. This is done
by extending EasyCurl to support multiple URLs. Kudu always attempts to
connect to the first one in the list, and if the connection fails due to
a network error or a timeout, it fails over to the next one. If a server
responds with an error, that is considered a valid response and then it
is returned immediately to the caller without trying the remaining URLs.

Change-Id: Ibef941ed20eda1f4e624c2c7e16ca7955af570d8
Reviewed-on: http://gerrit.cloudera.org:8080/19271
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>