[tablet] clean-up on Tablet::UpdateLastReadTime()
Tablet::UpdateLastReadTime() was marked 'const' while it wasn't so
semantically, but the comment in the header file was claiming it's
necessary.  However, it turned out that's not so as of now, and this
patch reduces the confusion.  I also took the liberty to make formatting
changes in the constructor to comply with the C++ code style guide.

This patch doesn't contain any functional changes.

This is a follow-up to ca957fb86736b1e07b18d80fdfa4e41c191e7072.

Change-Id: Iaf204d50bb3ee470790cefae983e6fb7b5a21c98
Reviewed-on: http://gerrit.cloudera.org:8080/19280
Tested-by: Kudu Jenkins
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>