KUDU-3425 fix flakiness in TestSnapshotScanBeforeAHM
This patch fixes the flakiness in the TestSnapshotScanBeforeAHM scenario
of the TabletHistoryGcITest test.  The root cause was in
IsAlterTableDone() RPC reporting 'all is done' prior the tablet replica
that served the scan request has received the update on the custom
config for the scanned table.

Setting replication factor to 1 from default 3 made the scenario stable.

before: 19 out of 1024 failed
  http://dist-test.cloudera.org/job?job_id=aserbin.1669681704.94942

after:   0 out of 1024 failed
  http://dist-test.cloudera.org/job?job_id=aserbin.1669686405.116123

Change-Id: Ib31e18c2a5ae68d05b4b0b0db15ff1cd00ac240b
Reviewed-on: http://gerrit.cloudera.org:8080/19284
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>