LookupRpc use the meta_cache's client
If kudu::internal::LookupRpc is consturcted it gets access to two
clients, one from the table and one from the meta_cache. They can be
different when the table was created with an other client. This can be
problematic if KuduSession::Apply(KuduWriteOperation*) is used because
the RPC timeout is extracted from the operation's table's client
instead of the session's client.

Change-Id: Ib5351b6d651af120b3f69a34ab1e40a05e9c535b
Reviewed-on: http://gerrit.cloudera.org:8080/19321
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>