Make HasColumn work without specifying col_schema
KuduSchema::HasColumn() determines whether a given column name is
present or not in a given schema. If the column is found, a new
KuduColumnSchema object is created, and it is assigned to the argument
col_schema pointer.

There can be cases, where we are only interested in the boolean return
value, and we don't want to obtain the actual column schema.
This patch enables to pass a nullptr as the col_schema argument, if
we don't want to obtain the column schema.

This came up, while working on another patch [1], where there are two
cases where I need to create column schema objects, although only the
boolean information is needed.

[1] https://gerrit.cloudera.org/#/c/19272/2/
    src/kudu/client/scan_token-test.cc

Change-Id: I736f55240157a676841db8428c84ee2d78b1cdcd
Reviewed-on: http://gerrit.cloudera.org:8080/19315
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>