[util] introduce WaitFor{Tcp,Udp}BindAtPort()
This patch introduces a new utility functions to be used in tests:
  * WaitForTcpBindAtPort()
  * WaitForUdpBindAtPort()

They are similar to already existing WaitFor{Tcp,Udp}Bind(), but have
semantics of waiting while some process binds to the specified port.
The difference that in these new cases the port number is known
beforehand, and it doesn't matter which process binds to the port.

This functionality will be used in a follow-up patch.

Change-Id: Ia466bf20040b35359d95fd66971e4af668d081e7
Reviewed-on: http://gerrit.cloudera.org:8080/19337
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>