[www] add slow scans section
We can get the history of completed scans from /scans page right now.
But it is not easy to distinguish slow ones among those.

I introduced an extra section to the scan page to show slow scans
separately. A scan is called 'slow' if it takes more time than defined
by --slow_scanner_threshold_ms.
The number of elements in the slow scans history is limited by
--slow_scan_history_count.

Change-Id: Ibd1dcac8b81c5eefd306e7020c9c52b3f28e603c
Reviewed-on: http://gerrit.cloudera.org:8080/19392
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>