[KUDU-3430] Implement getting flags filter logic in tool side
When using new version of Kudu tool (version: 1.16.0) to
execute command: "kudu tserver get_flags localhost:7050
-flags=block_cache_capacity_mb" on the old version of Kudu
TServer (version: 1.10.1)，it returned all flags not the
specified flags.

That because the filter logic was not implemented on the
server side (Kudu version 1.10.1).

So it is better to implement the filter logic on the Kudu
tool side also.

Change-Id: I0a2ddc2540d7aa70d3dd2e6c4101bb227e94c858
Reviewed-on: http://gerrit.cloudera.org:8080/19393
Reviewed-by: Yingchun Lai <acelyc1112009@gmail.com>
Tested-by: Yingchun Lai <acelyc1112009@gmail.com>