KUDU-3357 proxied RPCs for multi-master Kudu clusters
With this patch, it's now possible to run a multi-master Kudu cluster in
a local area network while processing RPC requests sent from external
networks.  The RPC requests are assumed to be forwarded by a TCP proxy
which advertises RPC endpoints for the masters and the tablet servers
of the multi-master Kudu cluster.

Refactored and unified test scenarios in client-proxied-rpc-test provide
the coverage for the newly added functionality.

This is a follow-up to 3f29b5da5f59ea96cfec0608226d5c35740884a6.

Change-Id: Id834d5fae8dcc2230de9f6b6429334c24052011a
Reviewed-on: http://gerrit.cloudera.org:8080/19404
Reviewed-by: Attila Bukor <abukor@apache.org>
Reviewed-by: Zoltan Chovan <zchovan@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>