[web] add maintenance op statistics information on web pages for data_retained_bytes
MaintenanceOpStats collects statistics about 'data_retained_bytes_',
which means amount of disk capacity in data blocks that would be freed if
current maintenence operation ran. But the protobuf file has no such
information, so it is not shown on web pages.

This patch adds a field in the protobuf file and updates the embedded web server
to show the retained bytes for a maintenance operation. Sample screenshot of
the maintenance manager page can be found here:
https://github.com/shenxingwuying/kudu/blob/dev/flyline/screenshot/data_retained.png

Change-Id: Iac8f0307705d57cea48901102b170c88d73d8c2e
Reviewed-on: http://gerrit.cloudera.org:8080/19407
Tested-by: Kudu Jenkins
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>