[tools] add --tables flag to 'local_replica delete'
This patch updates the 'local_replica delete' tool to allow deleting
multiple tablets by table name. Sometimes it is not convenient for
users to specify multiple tablet ids, but table names will be much
easier to get.

In order to support deleting all tablets of specified tables, the
<tablet_ids> argument is updated to <tablet_id_patterns> which
supports basic glob syntax.

The default value of --tables flag is empty, which means all tables
will be included, so the tool can achieve same effect as before if
this flag was not specified.

A new test is added to verify the new functionality.

Change-Id: I2d862a715f05179b2e8def0d1cdfe58c32299329
Reviewed-on: http://gerrit.cloudera.org:8080/19411
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Kudu Jenkins