KUDU-1945 Auto-Incrementing Column, C++ client
This patch adds the initial part of the cpp client side changes to the
auto incrementing column feature.

A new KuduColumnSpec called NonUniquePrimaryKey is added. Semantically
it behaves like PrimaryKey:
- only one column can have the NonUniquePrimaryKey ColumnSpec in a given
  KuduSchemaBuilder context,
- if it exists, it must be defined in the first place,
- compound keys are defined through a set function.
Functionally non-unique primary keys don't need to fulfill the
uniqueness constraint. An auto incrementing column is added in the
background automatically once a non-unique primary key is specified. The
non-unique keys and the auto incrementing columns together form the
effective primary key.

Some technical notes:
- The name of the auto incrementing column is hardcoded into the Schema
  class. This is a reserved column name, users can't create columns with
  it. On the client facing side, this reserved string is reachable
  through: KuduSchema::GetAutoIncrementingColumnName().
- In this initial version there is no support for UPSERT and
  UPSERT_IGNORE operations.

As suggested in the server side patch [1], a specific builder is added
to construct the column spec for the auto incrementing column.  Since
this pins down the properties of the column, I took the liberty to
remove unit tests and checks which verify these properties.

[1] https://gerrit.cloudera.org/#/c/19097/

Change-Id: Ic133e3d44cc56c8351e33d95b523ed7b6b13617b
Reviewed-on: http://gerrit.cloudera.org:8080/19272
Reviewed-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>