[clock] detect unsynchronized clock earlier
This patch adds an extra check for the return code of the ntp_adjtime()
call in SystemNtp::Init().  This is to allow for earlier detection of
unsynchronized local clock conditions if using the "system" time source.
Before this patch, if the local clock wasn't synchronized with NTP, that
fact would be detected later upon first call to WalltimeWithError().
The latter usually happens when a Kudu server tries to find current time
to generate a hybrid timestamp for an operation.

This is a follow-up to e72208436a625391739217394c67d783e992367a.

Change-Id: I32fd6ac1adfa04251e1e4b6237d650071e3b4dc9
Reviewed-on: http://gerrit.cloudera.org:8080/19449
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>
Reviewed-by: Yingchun Lai <acelyc1112009@gmail.com>
Tested-by: Alexey Serbin <alexey@apache.org>