[clock] output more info in SystemNtp::DumpDiagnostics()
This patch adds more details to the output of the DumpDiagnostics()
method for the system clock synchronized with NTP.  In particular,
the output now contains relevant metrics for the hybrid clock, the
output of the 'ntpstat' utility (if present), the output of
'chronyc activity' (if the 'chronyc' CLI is present), and a snapshot
of the newly introduced 'clock_ntp_status' metric.

As already mentioned, this patch also adds a new 'clock_ntp_status'
metric that provides information on ntp_gettime()/ntp_adjtime() call
result.  It's a string containing current timestamp (usec),
max error (usec), and the result status of the call.

The corresponding HybridClockTest.TestNtpDiagnostics scenario has been
updated correspondingly.

Change-Id: Ibb74e81ba1e4bc4c60acc6bd49bb458085e28207
Reviewed-on: http://gerrit.cloudera.org:8080/19478
Tested-by: Kudu Jenkins
Reviewed-by: Ashwani Raina <araina@cloudera.com>
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>