[Tool] Show the information of a tablet
This tool is used to show the table name and all replicas on
different tablet servers according to the tablet id.

Sometimes, the log prints the tablet id when a query failed or
the system got something wrong. But it is difficult to find
useful information according to the tablet id.

This tool will be helpful for debuging. The following is an
example:

Command: kudu tablet info $master_address $tablet_id

The result is:
 ID  | Table Name | Tablet Server UUID | Host | Port  | Is Leader
-----+------------+--------------------+------+-------+----
 abc | test_table |       xxxxx        | host1| 43689 | V
 abd | test_table |       xxxx1        | host2| 43035 | L
 dbe | test_table |       xxxx2        | host3| 35209 | V

Change-Id: Ib5ae5f61f50a44c4787843df76adaa61700ae9fe
Reviewed-on: http://gerrit.cloudera.org:8080/19501
Tested-by: Kudu Jenkins
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>