KUDU-1945 Auto-incrementing column feature flag
This patch adds the flag "--master_support_auto_incrementing_column" to
guard the auto-incrementing column feature. It prohibits users to create
table with non-unique primary key. The default value for the flag is set
to true in this patch. The verification happens on the RPC level. If the
server has the feature turned off, or the server is older compared to
the client - doesn't have the auto-incrementing feature -, the request
gets rejected in both cases.

Change-Id: I39c3dde3705c25c36d3ad787c0db6ed03f6c2cfd
Reviewed-on: http://gerrit.cloudera.org:8080/19523
Tested-by: Kudu Jenkins
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Reviewed-by: Alexey Serbin <alexey@apache.org>