[thirdparty] skip build/generation of LLVM benchmarks
This patch turns off generation and build of LLVM benchmarsk when
building LLVM in Kudu thirdparty.  In addition, it turns off the
generation of the Ocaml docs.

Doing so allows for building LLVM on Ubuntu 22.04 without patching
the LLVM code itself.  It also speeds up the build just a bit.

Change-Id: I67592101c5e105b8a4b36fc231b0e67e4585fb0b
Reviewed-on: http://gerrit.cloudera.org:8080/19514
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>
Tested-by: Yifan Zhang <chinazhangyifan@163.com>