[fs] update on FsManager's constructor
While reviewing on of the recent patches, I noticed that the signature
of FsManager's constructors might be improved to avoid specifying
default instance of FsManagerOpts as the second parameter.  This patch
addresses that and also cleans up instantiation of FsManagerOpts in a
couple of places in kudu CLI tools.

This patch doesn't contain any functional modifications.

Change-Id: Ib76723a230142a4903cca4f47fbde44afa15abed
Reviewed-on: http://gerrit.cloudera.org:8080/19570
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>
Tested-by: Yingchun Lai <laiyingchun@apache.org>