[tools] exit gracefully on errors during table copying
Prior to this patch, an error to write a batch of rows into the
destination table would lead to a crash while running `kudu table copy`
CLI tool.  In addition, the information on errors would not be printed
because CheckPendingErrors() was called before Session::Flush().

This patch updates addresses these issues:
  * no crashes for non-OK status within TableScanner::CopyTask()
  * information on errors encountered during Session::Flush()
    is printed into the output stream

This is a follow-up to 28f8e972fcd042a26a502cbb1f1102c487c9398d.

Change-Id: I90dc29b5a3fb8334ec1f54425b84d13faeb19cd5
Reviewed-on: http://gerrit.cloudera.org:8080/19575
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>