KUDU-3437 Set default block_cache_metrics_policy to reset
If the master has empty local directories and connects to an existing
cluster, it executes Master::Init twice, which executes
StartInstrumentation twice, which is only allowed if it's a test or the
policy is to reset the existing CacheMetrics. According to documentation
ExistingMetricsPolicy::kKeep should only be used during tests, so the
default value is set to kReset which fixes the original issue.

Change-Id: Idc32d7ab02201382debcbe36311579550353bf71
Reviewed-on: http://gerrit.cloudera.org:8080/19585
Tested-by: Kudu Jenkins
Reviewed-by: Yuqi Du <shenxingwuying@gmail.com>
Reviewed-by: Alexey Serbin <alexey@apache.org>
(cherry picked from commit e07c941aa6622efb758283cb2a583d0b90deb830)
Reviewed-on: http://gerrit.cloudera.org:8080/19673
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>